Upgrade tauri to v2
