+++ package.json
@@ -14,11 +14,21 @@
   "author": "Ajay Bura",
   "license": "AGPL-3.0-only",
   "dependencies": {
-    "@tauri-apps/api": "1.6.0"
+    "@tauri-apps/api": "2.7.0",
+    "@tauri-apps/plugin-clipboard-manager": "2.3.0",
+    "@tauri-apps/plugin-dialog": "2.3.2",
+    "@tauri-apps/plugin-fs": "2.4.1",
+    "@tauri-apps/plugin-global-shortcut": "2.3.0",
+    "@tauri-apps/plugin-http": "2.5.1",
+    "@tauri-apps/plugin-notification": "2.3.0",
+    "@tauri-apps/plugin-os": "2.3.0",
+    "@tauri-apps/plugin-process": "2.3.0",
+    "@tauri-apps/plugin-shell": "2.3.0",
+    "@tauri-apps/plugin-updater": "2.9.0"
   },
   "devDependencies": {
     "@actions/github": "6.0.0",
-    "@tauri-apps/cli": "1.6.2",
+    "@tauri-apps/cli": "2.7.1",
     "node-fetch": "3.3.2"
   }
 }
